Doc: QtCore: Add missing \headerfile topics
A lot of functions and macros previously available under the QtGlobal
header have been separated into different, specialized headers. The
documentation for these were updated to refer to the new header file
names using the \relates command, but a \headerfile topic was missing
for a lot of them.

As a consequence, QDoc automatically generated a 'proxy' page for these,
which look confusing as they lack any description of what the page is.

Add \headerfile topics for them, along with \brief descriptions.

Fix \relates arguments for type alias QPair and the deprecated function
qMakePair() to point to the correct header file.

Pick-to: 6.10.0 6.10 6.9 6.8
Task-number: QTBUG-117447
Change-Id: I98114ddb71f78982e390e28e9521e264f3ecf5d0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>