QRM: add test for headerData, remove storage specifiers and wrappers
The test was missing, now it isn't. Check only for the horizontal header
where QRM provides a default implementation for cetain types, and and
the first column which tells us enough.

Fix failing tests: we don't want to include the storage specifier or
smart-pointer wrapper types that might be used to control the lifetime
of objects in a tuple.

Pick-to: 6.10
Change-Id: If35abff3d8350c5402e8bbc0519a12420d4b6d8d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>