QMacStyle: protect against a nullptr context
Found in the stylesheet auto-test that started crashing after a recent
change (introducing a manual slider painting). The test is using QImage
with format QImage::Format_ARGB32 to render to and CGBitmapCreateContext
fails for such format. Previously was not a problem, since we did not
use the context directly.

Protect against nullptr in the NSView drawing function and in functions
like drawControl etc. (but do not skip the parts using QPainter without
triggering NSControl/CGContextRef-related code-paths).

Fixes: QTBUG-139983
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Iee1aa23549b29b489b870be3c7660700eba2cbc2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>