qnumeric.h: add support for C23/C++26 <stdckdint.h>
This allows us to have cross-platform support instead of relying on
intrinsics. Though this requires banning 'char' because this type
changes signedness I guess and C23 doesn't allow them.

Both Clang and GCC have a C23 header for <stdckdint.h>, added for GCC 14
and LLVM 19. Both sets of headers work in earlier C versions. Neither
MSVC nor MS STL have added the headers, but once they do we should get
the proper constexpr and runtime-optimal content that they came close to
adding[1], where I reported:

> For my own codebase, I don’t plan on writing new code. Users of Visual
> Studio will keep getting worse codegen for a few more years, unless
> they switch to Clang or GCC.

GCC's header does not work in C++ due to the use of _Bool. That is fixed
by libstdc++ from GCC15, with a C++ <stdckdint.h>, so we require that
version of GCC. This will fail to compile with GCC and libc++ until the
bug[2] is fixed.

[ChangeLog][Source-Incompatible Changes] qnumeric.h's qAddOverflow(),
qSubOverFlow(), and qMulOverflow() functions no longer accept char as
a template parameter type. Use quint8 or qint8.

[1] https://developercommunity.visualstudio.com/t/please-implement-integer-overflow-detection/409051
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=121811

Change-Id: I5ca93cf4166735075b1bfffdad1dcd7c20cab031
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>