QVariant: delegate to move-constructor in the QMetaType constructor
Instead of the move-assignment operator. The compiler is apparently not
smart enough and does create an extra QVariant from which it copies and
must then destroy. In fact, there were two calls to ~QVariant.

Now there are none and no 32-byte copy. It's effectively a tail-call
jump to QVariant::fromMetaType().

Pick-to: 6.10 6.9 6.8
Change-Id: I483b4273b28ced967c29fffd54857fff69db66e5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>