Android: fix rendering glitch after window size changes
Currently, apps might end up with half of the window size not rendered
or not reflecting size changes. When the root layout size changes, the
available size is reported but the screen size is not always sent from
Android callbacks, this is the case after an orientation change. So
make sure here to update the the screen geometry also when receiving
the available geometry.

This part amends 7dcf58eedbeb0cbfbf01f3aabfb72f8546f96cdf.

However, the issue was still present in 6.9 and 6.8 before that above
commit, to fix that, it's needed to call sendExpose() to send expose
event for the window in the new region, and this after propagating
the surface change from QtSurface to the platform window.

Pick-to: 6.8 6.9 6.10 6.10.0
Fixes: QTBUG-132718
Fixes: QTBUG-134082
Task-number: QTBUG-124140
Change-Id: I390aaba45f324693dc47d7d0110fdb22dc9aca61
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>