QVariant: fix GCC -Warray-bounds in rvalue qvariant_cast
Change prompted by this warning from GCC:

  In function ‘T qvariant_cast(QVariant&&) [with T = QTransform]’,
      inlined from ‘T QVariant::value() && [with T = QTransform]’ at qvariant.h:530:30,
      inlined from ‘void QSvgAnimatedStyle::applyPropertyAnimation(QPainter*, QSvgAbstractAnimatedProperty*, bool, QSvgExtraStates&)’ at qsvgstyle.cpp:754:87:
  qvariant.h:794:67: error: array subscript ‘QTransform[0]’ is partly outside array bounds of ‘QVariant [1]’ [-Werror=array-bounds=]
    794 |             return std::move(*reinterpret_cast<T *>(v.d.data.data));
        |                                                                   ^
  qsvgstyle.cpp: In member function ‘void QSvgAnimatedStyle::applyPropertyAnimation(QPainter*, QSvgAbstractAnimatedProperty*, bool, QSvgExtraStates&)’:
  qsvgstyle.cpp:754:67: note: object ‘<anonymous>’ of size 32
    754 |         QTransform animatedTransform = property->interpolatedValue().value<QTransform>();
        |                                        ~~~~~~~~~~~~~~~~~~~~~~~~~~~^~

If sizeof(T) > sizeof(Private::data), the sought object cannot
possibly be in the internal storage. But the old code still checked
is_shared and exposed the std::move(Private::data) case to the
compiler, prompting said warning.

To fix the warning, and to avoid emitting the redundant runtime check,
first check whether the sought object could possibly fit into the
internal buffer. Don't use CanUseInternalBuffer, which is what the
original implementation¹ did and we fixed afterwards². Just do a size
check. If the object is too large, no manner of QTypeInfo flags will
ever make it use the internal space. Since this is a compile-time
check, wrap it in a constexpr-if, thus hiding the move-from-internal-
space from the compiler for objects that are too large.

Add some checks, also for get/get_if. I couldn't reproduce the
warning, so I don't know whether these tests threw them, but the
setting is the same as in Eddy's report.

¹ dcf76042301772b770a088526e5753ed41232d87
² eb87b0444ac8fec4d86bc26dc4bec80f82953f7c

Amends [²].

Reported-by: Edward Welbourne <edward.welbourne@qt.io>
Fixes: QTBUG-140064
Pick-to: 6.10 6.10.0 6.9 6.8
Change-Id: I66c12506f1746a94138c65dad980160138ab0a9e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>