CMake: Rework SBOM version handling
The SBOM project version handling had some issues.

If USE_GIT_VERSION was used by a project, the spdx file name would
never get a version suffix, even if the project also passed an
explicit VERSION option.

So either USE_GIT_VERSION or VERSION was used, but not both.

Rework the code so that first the git version if is queried if
USE_GIT_VERSION is passed, and then that gets overridden with the
value of VERSION if that is passed.

VERSION will be used as the version suffix for the spdx file
name even when git info is available, by storing it in the
QT_SBOM_EXPLICIT_VERSION global property and using it at the
appropriate call site.

Also add a new QT_SBOM_VERSION_OVERRIDE variable that can be used
to override the explicit version, if needed.

Pick-to: 6.8 6.9 6.10
Task-number: QTBUG-134894
Change-Id: I999f731493ddffcc6974edeca0adacefcc85b9d9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>