QLocale: de-inline c()
At face value, this is just replacing one out-of-line function call
(the QLocale(Language) ctor) with another (QLocale::c()). It follows
that this isn't worse than what was there before.

But as an out-of-line function, it has access to c_private() directly,
and we can afford to commit to noexcept(true) in this special case,
because we'll always be able to create this special object in a
noexcept way, so add that, too.

Together with the fact that the ctor we've been calling from inline
code doesn't just take one argument, but three (although the defaulted
ones are all of trivial type, and the default values are all
numerically zero), and this static function now takes zero, this
should nicely reduce the code generated for calls to this function.

Change-Id: If6ec4205773a57f1b3d3acf55d10ad4e030d92b8
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>