Mark QChar as security-critical
QString and QByteArray are critical, too, and not because of their
ownership semantics, but because of the algorithms operating on
them. QChar contains at least decomposition() as one of those
algorithms.

The implementation in the header is just a bit too much to be called
trivial-impl-only, so mark the header critical as well.

Task-nunber: QTBUG-135195
Pick-to: 6.10.0 6.10 6.9 6.8
Change-Id: I58aac5e290fa9487340b8df500b915b3b150e38a
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>