Move CharBuff back into QLocaleData
This reverts commit 69d3e96e02d4726d79aaf2d82acd30a903aa5a74, and
adjusted users that have since been added.

Its name doesn't start with Q so it shouldn't be in the public
namespace, even if it is in a private header.

The reason for the move out of QLocaleData disappeared with the
recent commit 1cdb5178b9db59f3aa5c0af8f30ad25bc6620021, which moved
ParsingResult into QLocaleData for the same reason.

Pick-to: 6.10 6.9 6.8
Change-Id: Icd3eac41ef7654024224dd1d55a5922f7f598a36
Reviewed-by: Marc Mutz <marc.mutz@qt.io>