Android: don't recreate a new ANativeWindow on size change
When only the surface's size changes we shouldn't recreate a new
ANativeWindow but only handle expose event with the new region
and keep using the existing android native surface.

At the same time set the ViewGroup.LayoutParams to the native
window/view to avoid any unnecessary calls to setGeometry() when
a resize happens. So instead of setting a specific size and have
Qt call setGeometry when the root layout resizes then we end up
setting new sizes again for QtWindow and QtSurface, we will have
a simpler cycle where the root layout changes and in the same
pass it would resize it's children on the Android side, we then
get the size change callbacks, then update the Qt window's size,
then even if Qt ends up calling setGeometry on the native side
again it most likely won't do much because the View already has
that requested geometry.

Pick-to: 6.11 6.10 6.8
Fixes: QTBUG-132721
Fixes: QTBUG-132717
Fixes: QTBUG-132720
Change-Id: I2253e81ffe6c69687cab5279572eacb26b5436b7
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>