Mark qlocale_data_p.h as security-critical
While the vast majority of the file is generated tables (which
shouldn't be in a header, cf. QTBUG-128930), there's AlphaCode, which
has an open bug potentially exposed to external input
(cf. QTBUG-138562), so clearly, while that component is in the header
(a patch in the chain for QTBUG-138562 moves it out of there), the
header is security-critical.

Task-number: QTBUG-135195
Pick-to: 6.10 6.10.0 6.9 6.8
Change-Id: I866938c4bc7bc5be4cce70f04b91fbc1a8d13a4e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>