QTextBoundaryFinder: add missing move SMFs/swap()
The class was (deeply) copyable, but not movable. Add the missing move
SMFs, member-swap(), and Q_DECLARE_SHARED (which adds ADL swap() and
QTypeInfo).

The inline operations do not break encapsulation, and they bind future
extensions only insofar as the freeBuffer variable only gets copied;
the source isn't modified. This is ok for flags, incl. for freeBuffer
itself, as free(nullptr) is a well-defined no-op. If we need more,
chances are we wouldn't have fit it into the 31-bits "unused" bits,
either, and we would need a V2 class, anyway.

Fixes: QTBUG-138659
Change-Id: Idea621b8cc8e4fce471071099f70d3ab9d20f368
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>