Inline NumericTokenizer::asBmpDigit() and rename to digitValue()
The old name suggested it would return something in the BMP, thus some
form of character, but it returns the digit value, so name it to match
the NumericData method it now packages. That packaging now means it
can in fact take a char32_t, so give it that as signature. Inline and,
in the process, remove the spurious inline keyword from an adjacent
method with inline body.

Rework the callers so that we now use it consistently.

Change-Id: If879173d1872f4cf14de32a304d86af7a8f330d6
Reviewed-by: Mate Barany <mate.barany@qt.io>