QToolButton: always set to MenuButtonPopup when action has a menu
The old implementation only set MenuButtonPopup when the corresponding
QAction had a menu *before* it was assigned to a QToolButton but not
afterwards which leads to inconsistencies. Make sure to always set
MenuButtonPopup when the QAction gets a menu. Also mark the todo from
Qt4.2 with a proper Qt7 todo marker so it can be found and fixed during
Qt7 transistion.

Pick-to: 6.10
Task-number: QTBUG-140148
Change-Id: I4def22e7147049c30a16a7fe88c18591dbb58e88
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>