QTextBoundaryFinder: eliminate two 4-byte padding holes in Qt 7
By reordering the members smaller than a pointer to the end. Amends
commit 0ae5b8af9c2af514f316211b71a03defc0f1b65a, which changed the pos
field from int to qsizetype, though a padding hole existed before that
commit too. The correct ordering would be to put qsizetype pos after the
attributes pointer, since qsizetype can be reduced in size to int, but
that's unsupported and I can't be bothered.

There's currently a 3-byte padding between the new boolean field and
the BoundaryType. And as a bool type, it is using all the 8 bits in
the member. The padding is fixed in the next commit, explicitly
marking the reserved space.

Drive-by apply NSDMI to the freeBuffer, which we couldn't before commit
cc16e9e4030f310cf344b340a01c99c257f3841b because it was a bitfield.

Drive-by fix coding style of the member initialization list.

Change-Id: I1701f4a73b7810dc6efafffd4ed34b822c36cac3
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>