Android: Fix double emission of certain a11y events
3c709198838866d5122c69a30cacdc806605d0cf and
e7a211d702d6aa21a30eff56ceeffc153385443e introduced a11y actions for
focusing and scrolling. However they also accidentally introduced a
double emission of their according events. This causes problems, since
every new event stops the processing and announcement of the previous in
TalkBack.

We remove the sendEventForVirtualViewId() calls from performAction() and
performActionForVirtualViewId(), because they are the culprit. If an
action like focus or scroll changes the app state, an a11y update is
triggered, which ends up calling notifyObjectFocus() or
notifyScrolledEvent() where the according event is then emitted and the
a11y delegate state is updated.

For the focus there is an exception, since an element cannot gain the
active focus if is is disabled. However TalkBack should still be able to
place its focus frame on that element and read out its content.

Pick-to: 6.10 6.8
Change-Id: I9061eec713fd5382feb7aacafdffe393968ef8f1
Reviewed-by: Lars Schmertmann <SmallLars@t-online.de>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>