Android: Avoid announcing simple text as clickable
The definition of Accessible.onPressAction is enough to let
Android announce an element as clickable. But this is a
problem when components are designied in a way like this:
  Accessible.role: clickable ? Accessible.Button : Accessible.StaticText
  Accessible.onPressAction: clickable ? doit() : nothing()
So we avoid to set clickable for StaticText and Heading.
Windows, macOS and iOS already have the right behavior and do
not care about pressAction or toggleAction on simple text.

Pick-to: 6.10 6.9 6.8
Change-Id: I7bf9ffae112b92dd4a90f51f014a348dfe1b63f3
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>