QAppleNetworkInformationBackend: fix races, protect against a dangling pointer
nw_path_monitor's update handler calls our 'updateState' on a dispatch
queue, potentially from a different thread. 'this' pointer captured
by the handler's lambda can be already invalid (if the backend was deleted).
We call nw_path_monitor_cancel in 'stopMonitoring', but it's not enough
to prevent update handler from accessing the (now-invalid) pointer.
To resolve this issue we move the lock and 'this' pointer to the separate
object, which is essentially owned by the callback and will be deleted
by the monitor's cancellation handler, set via
nw_path_monitor_set_cancel_handler.

Pick-to: 6.10
Change-Id: I79f48ef021d19dcba0923ef06690433d167c4ea6
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>