QDate: make weakly incrementable
It is logical that incrementing a date by 1 results in the following
day (rather than the following month or year). Add the pre- and postfix
increment operators as hidden friends, and declare a difference_type
to be qint64. With these additions, QDate models the
weakly_incrementable concept, allowing for it to be used with
std::views::iota.

For symmetry, although not needed for the weakly_incrementable concept,
also add operator--. One could add operator+/+= for std::chono::days,
but it is neither adding symmetry, nor is it required for a relevant
concept, so not doing that (in this commit).

For QTime (and from that, QDateTime), there is no logical "next";
incrementing by a Planck time (as the smallest unit of time) is not
useful, and anything else is arbitrary.

[ChangeLog][Core][QDate] QDate now models the std::weakly_incrementable
concept by implementing pre- and postfix increment (and decrement)
operators.

Change-Id: I87e68f014834446b6444f23ff0e2f16bdc227a14
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>