QMacStyle: set NSStepperCell's control size to 'mini'
On macOS 26 this thing became rather large, not fitting into the
simple clipping rectangle of a spinbox (in a native app such a stepper is
taller than its controlled input, e.g. date/time edit). In Qt, with clipping
applied, this looks not so nice; but by always setting the cell's control size
to 'mini', we can have a properly looking spinbox, similar to what it was before.

Fixes: QTBUG-140134
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I4c3798e84e602aceead5770963fde66e9c106e53
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>