wayland: Ignore wl_pointer.leave serial
That serial is not suitable for anything useful. It cannot be used for
interactive move/resize, setting a selection, or any other requests.
It is also not suitable for event ordering. Qt doesn't use these serials
itself, they are mostly useful for communicating back to the compositor.

The serials in wl_keyboard.leave events are already ignored. So it also
improves code consistency.

Change-Id: Ieb684c369ae665c4aeb08a6dc4be31206e0b6ed0
Reviewed-by: David Edmundson <davidedmundson@kde.org>