Android: change only the opacity of system bars when setting system ui
When switching between expanded mode (similar to edge-to-edge) and
back to normal mode, we shouldn't be touching the colors of the
system bars, as that can mess up with the theme values or user
defined values. Instead, simply change the opacity (alpha bits)
of the color so we get transparent bars in expanded mode and
opaque bars in normal mode.

Task-number: QTBUG-139690
Pick-to: 6.10 6.8
Change-Id: Iadc5a9d8625c50009a0dc1a5ac29b5354089d345
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>