QWidgetBaselineTest: Draw all sub controls, at right position
In some cases a style's drawComplexControl won't exercise the entire
range of subControlRects, or won't do so for each widget if multiple
instances of the same widget are drawn, so we opt to resolve the
sub control rects manually for calls to drawComplexControl. As part
of that we make sure to map the subcontrol rect to the rect of the
control.

Change-Id: Iaca1dcee1a0141e6dc8c81a6b8a4d3ed9be195e6
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>