QLockFile/Win: replace DeleteFile with CreateFile w/o sharing
What looks like an antithetical change actually makes sense: we open the
file with FILE_FLAG_DELETE_ON_CLOSE, which causes the file to be deleted
when we CloseHandle(). In fact, this is exactly how wine implements
DeleteFile[1].

The difference between this and DeleteFile itself is that we are opening
the file for writing but not allowing any sharing. This means no other
process in the system can open the file at the same time as we do, even
if they want to cooperate in deleting it.

[1] https://github.com/wine-mirror/wine/blob/wine-10.15/dlls/kernelbase/file.c#L1011

Fixes: QTBUG-140053
Pick-to: 6.10 6.9 6.8
Change-Id: Iba265eba236f496ce5c9fffdc5daf95d5f429358
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>