Android: call view's onApplyWindowInsets() in QtInputDelegate
This callback was done on the decor view and  wasn't calling the base
onApplyWindowInsets() practically ignoring base insets like system
decorations, this then messes with the default behavior of system
bars making calls to customize color schemes have unexpected
behavior at least or at worse not working at all. Fixing this brings
back the expected behavior of the insets when in the case of the
linked ticket, we would get back correct handling of navigation
bar semi-transparent bars where the bar icons have low contrast
compared to the content below it.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-139690
Change-Id: I5b46f6f6e0c7850fba117a0ba76d3a693c1cb37b
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>