wasm: fix memory corruption in Promise handling
The "finally" handler deregisters and deletes itself, but
would do so before it was completed. Move the removeEventHandler()
call to the end of the lambda function.

As it happens this did not cause any issues on wasm32,
but for wasm64 finallyFunc would become corrupted and
crash with an "null function or function signature mismatch"
error.

Pick-top: 6.10 6.9.3
Change-Id: I66125190677843593866b8eec950ac60d2bb75ce
Reviewed-by: Even Oscar Andersen <even.oscar.andersen@qt.io>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>