QRM: don't leave gaps in role values for non-standard properties
When mapping properties as role names to role values, we use the pre-
defined values for properties that match the standard role names, e.g.
a property "display" is mapped to Qt::DisplayRole. Properties with
custom names are then mapped to Qt::UserRole...

If we leave gaps in the Qt::UserRole... sequence when a property with a
standard name is found, then the role values become unpredictable and
hard to use correctly from code. So don't increment the role value when
a property with a standard name is found.

Add test coverage.

Pick-to: 6.10 6.10.0
Change-Id: If54833e94aa52ee5f387029ca4c88ad6b9d2348e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>