QRM: don't write to an invalid property
The role value might not map to a property, in which case the
QMetaProperty will be invalid. We already handle this in readRole, so
test for the valid property in writeRole as well.

Add crash-test to the setData() test case; we can write arbitrary role
values to items backed by maps, so don't verify the return value of
setData.

Pick-to: 6.10 6.10.0
Change-Id: I780281cb7f1bbbe6db1f66dcc978ae45a443b716
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>