QRM: add customization point for item access
While using gadgets and QObjects with property access through the meta
type system is convenient, it's also somewhat slow and inflexible. The
order of properties defines how they get mapped to Qt::UserRole values,
and even with the cache for the mapping of role to QMetaProperty, the
data access goes through a number of lookups.

Add an ItemAccess nested template class that type authors can specialize
for their own item type to implement optimized read and write access for
the roles their item type supports.

The benchmark shows that read access is twice as fast as going through
the meta property system for an otherwise identical gadget type, and 30%
faster for write access.

Change-Id: I9b1b101cb477a7853d5fe0378c624f8459385da6
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>