Check unset strings are null in QLocaleData::NumericData tests
Verify that the strings are null, rather than merely equal to the
empty strings passed, when they should not have been set at all.
This incidentally means the string data passed for the C locale get
ignored, as the zero already is, so passing their nominal values
instead of empty strings is harmless, perhaps more informative and
more consistent with other locales' data.

Change-Id: I1bfa76ce0ac7365669bf482b7e1fff9a1a25030a
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>