Allow binding a socket to a specific QNetworkInterface
For now, via a private API.

Provide a new socket engine option - BindInterfaceIndex, and use it if
a valid interface is passed to QAbstractSocketPrivate::bind().

Note that this feature is not fully functional on macOS:
both SO_BINDTODEVICE and IP{V6}_BOUND_IF only work for limiting
outgoing datagrams. The patch uses IP{V6}_BOUND_IF, because it allows
to avoid number -> string -> number conversion.
As of now, extra filtering of incoming data should be done on the user
side.

Task-number: QTBUG-80704
Task-number: QTBUG-139697
Pick-to: 6.10 6.8
Change-Id: Ic207908313d9d25f96c23ecc363181ff0ae8232a
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>