Android: make system ui visisibility flat with no cached values
If system ui visibility is requested simply send the event and
let the system decide what to do about it, that way we don't
need to keep any cached values. Also, when checking for states
like fullscreen and expanded modes figure that out from current
system ui state instead of fetching the cached values which is
more accurate because the system ui is usually not updated
immediately.

Change-Id: Id186d69a0e338e9533d58a23f1bd2b74612d9a88
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>