Mark qdoublescanprint_p.h as security-critical
While the implementations are one-liners, they cannot be said to be
trivial, so this string → floating-point parser can't be
significant-only as trivial-impl-only.

Another option would be to say that the callee firewalls this component
from security issues, like 3ba8e2fbc483f8f5564bb30ee7437a9150e193b5 did
for QStringList/QString. But this component is adding requirements of
its own on top of the requirements of the callee (e.g. correct use of
the QT_CLOCALE* macros), so claiming the callee firewalls it off would
be a tall order.

In the author's opinion it must therefore be considered critical
instead.

QUIP: 23
Task-number: QTBUG-135195
Pick-to: 6.10 6.8
Change-Id: I167b821f0b1765d6b3a1d97fcd0f9edae5243778
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>