Add private QIconEngine hook to get underlying platform icon
This allows us to pass the underlying NSImage from the Apple icon engine
to the NSMenuItem.image, without rasterizing it ourselves in the process.

The latter is both inefficient, and does not result in correctly sized
or aligned menu icons, as we don't have enough information at the point
when we rasterize the icon in QAppleIconEngine::paint to determine the
correct NSImageSymbolConfiguration for a menu icon (NSFontWeightSemibold,
NSImageSymbolScaleSmall, and [NSFont menuFontOfSize:0].pointSize).

By passing on the underlying NSImage we fix the broken icon sizes in
native menus, but we still have wrongly sized icons in non-native
menus (context menus in Qt Widgets, and non-native menus in Qt Quick).
This comes down to the pixel metric for icon sizes (PM_SmallIconSize)
being built for square icons, which does not work for the dynamic
nature of the SF Symbols icons. In addition, PM_SmallIconSize is
used for more than just menu icons, so we can't just update its
value to make the non-native menus look slightly better.

Fixes: QTBUG-140361
Pick-to: 6.10 6.8
Change-Id: I1e2cb4029688c824ec29546836a0820cb635722c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>