a11y: Simplify/Deduplicate qAccessibleTextBoundaryHelper
Extract a local helper function qAccessibleTextBoundaryHelperHelper
that takes care of moving the cursor to the start
and end positions and returns the corresponding indices.

Simplify and unify QAccessible::qAccessibleTextBoundaryHelper
to use that helper for multiple cases instead of repeatedly
implementing equivalent logic. Also return the result early
instead of first assigning it to a local variable.

No change in behavior intended.

Task-number: QTBUG-140504
Change-Id: I10df3957166cf815220a3c67f1774fc37c5e7311
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>