a11y: Extract helpers for getting text at/before/after offset
Move the existing QAccessibleTextWidget logic to implement
the QAccessibleTextInterface::textAfterOffset,
QAccessibleTextInterface::textBeforeOffset and
QAccessibleTextInterface::textAtOffset interface
methods using a QTextCursor out of that class to
3 new QtGuiPrivate helper methods.

This will allow to reuse them in QAccessibleQuickItem
in qtdeclarative, which implements the exact same logic
(for one code path).

In a first step, only move the existing logic, i.e.
no change in behavior is intended by this commit.

In an upcoming commit, the actual logic will be improved,
which will then fix both, the QAccessibleTextWidget
and QAccessibleQuickItem implementations without having
to keep the code in sync manually.

Task-number: QTBUG-140504
Change-Id: I87690fb6af040d6c9207bf19ddffe2e3eba16b0d
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>