a11y: Fix cursor-based text{Before,After}Offset
So far, qt_accTextBeforeOffsetHelper and
qt_accTextAfterOffsetHelper (used by
QAccessibleTextWidget and QAccessibleQuickItem)
were not consistently returning text items of
the given boundary type that is completely
*before* or *after* the given index, but
could e.g. return the word in whose
middle the offset is when using
QAccessible::WordBoundary.

This could for example break an assistive technology
feature of navigating through a text by word
if it relies on those methods.

Adjust the implementations to prevent that
from happening, by calling
QAccessible::qAccessibleTextBoundaryHelper
again with an index moved by one character
until either a result before/after the given
index is returned or the start/end of the text
has been reached.

Extend the existing QTextEdit a11y test
by a few cases that would previously fail.

(Note: Depending on the text granularity used,
not every character in the text is necessarily
considered part of any text item: E.g., when using
QAccessible::WordBoundary, space characters in between
words are not, at least not for the QTextCursor logic
as used in QTextEdit or qtdeclarative's QQuickTextEdit).

Fixes: QTBUG-140504
Task-number: QTBUG-139943
Change-Id: Ie6f0f2b2579237f42964d2498c818d1f02bf4bb8
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>