QtCore/time docs: give \overload its parameter
Hopefully this will give a better user experience for those reading
the docs online. Exploits a feature new in dev (6.11) so no picking:
the primary overload is now marked with \overload primary. I've also
added missing \overload to a few functions that were overloads and
moved one stray comment from between a qdoc-comment and its function.

In the process, impose some consistency on ordering of commands:
* \since before \overload before other things
* always indent four spaces (some \overload were just two)
* always separate actual text from initial commands with a blank line.

Fixes: QTBUG-140485
Change-Id: Ib8550051c40efa42240d96b2d14e3ba746790f0c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>