QWindows11Style: Handle inverted sliders correctly
QSliders with `setInvertedAppearance(true)` didn't show the blue groove
and ticks correctly. With this change, the highlighted groove is on the
right side if a hozizontal slider is inverted and on the top side if a
vertical slider is inverted.
Note that `upsideDown` is true for non-inverted vertical sliders but
false for non-inverted horizontal sliders.

Pick-to: 6.10
Change-Id: I8c75dbc1006349209316fce58a5ec383221a916f
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>