Add security headers for src\printsupport
As defined in QUIP 23, security critical files should be marked with
Qt-Security score:critical. This change also marks all non-critical
files as significant (the default) to make it clear that all files
have been considered.

As all printing to the device using the input data falls under the
responsibility of other modules, such as parsing a document and painting
it to the device. There are no direct network access or parsing
operations in the printing subsystem other than wrapping around the
native APIs.

Then we cannot mark anything here as Critical.

There are some parsing functions like QPrintUtils::parsePpdResolution,
QCUPSSupport::findPpdOption, and QCocoaPrintDevice::defaultColorMode(),
but all rely on the native ppd APIs.

Fixes: QTBUG-134646
Pick-to: 6.10 6.9 6.8
Change-Id: Ic02a81051f4a03f9caf188f091dcbf29f8094ed1
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>