tst_QLocale::long_long_conversion_data(): modernise and tidy
Change row names to use a / as separator, instead of a space, between
locale and text parsed, so that leading and trailing space can be
indicated by simply appearing as such, without getting confused with
the separator, instead of needing runic annotations to allude to them.

Use modern-style string constants, use C++-style casts.
Use an alias for qlonglong for the sake of brevity.
Purge space between parameters (that tried to create a table format).
Split the single newRow() line that still ran past column 100.

Change-Id: I3232d85c633ef85b3510611b1218cbbd18ee3734
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>