Add more test-cases to tst_QLocale::long_long_conversion()
These cover other character sets than Latin and explore the
complications with signs that include BiDi marks, and similar
invisible Unicode markings. Also refine some comments describing test
rows of other tests, where I now understand what they really test.

It turns out that, because we special-case ASCII digits and the U+2212
MINUS SIGN, merely omitting invisible decorations to signs works as it
should, but if any are present they have to match exactly. Include two
test cases that omit the (redundant) second left-to-right marker on
the signs of Punjabi in the Arabic script; this, as anticipated,
exhibits the expected bug by failing to be relaxed about handling of
such invisible marks.

That is somewhat of an obscure corner-case, that turns out to be
rather expensive to fix (incurring an O(20%) slow-down to all
number-parsing), so I'm leaning towards leaving it unfixed unless I
can find a more efficient approach to fixing it. But the test-case may
be some help if we ever do see such cases reported in the wild, if
only by leading someone back to the Jira ticket.

Task-number: QTBUG-139922
Change-Id: Ief159910d49342a0776cd8920ac8ce1d80280164
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>