iconbrowser test: add theme loading support
Add the root path to the search paths, which makes an absolute path to
a theme file load that theme. Move the UI into a main window and add
a file menu to allow loading of such a theme at runtime.

This shows that switching themes will reload the icons, at least in
widget UIs. Add two such themes with a single icon for test purposes.

Change-Id: I181fe4ac43a125b8f15fd38d302c4a4f7dc4dc6c
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>