QMacStyle: Draw disclosure triangles according to correct direction
When using a QTreeView and overriding its' (or any ancestor widget's)
layout direction from the default, on the Mac style the indicator
branches would be rendered pointing the wrong way - as those are drawn
as native disclosure triangle buttons whose NSCell did not get the
proper layout direction passed on to. This fixes it.

Pick-to: 6.10 6.9 6.8
Change-Id: Ic708e82dd2534af674822fa2432ce4a4af22c522
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>