CMake: Handle internal modules in qt_internal_wrap_private_modules
If a Qt module A linked against an internal module B privately, we
accidentally find_package()'d B when a user project looks for A.

Such a case happened with QtLocation which links to
QuickShapesPrivate.

This happens because qt_internal_wrap_private_modules only wraps
dependencies with $<BUILD_INTERFACE> if the dependency has the
_qt_is_private_module property set, which internal modules don't have.
This means we passed the internal module to
qt_internal_register_target_dependencies which then adds a
find_dependency() call for it.

Add code to check for the _qt_is_internal_module property and wrap
such a target dependency in $<BUILD_INTERFACE>, but only if it the
consuming target is not a STATIC library.
In the STATIC case, adding the genex would break user builds,
because the internal module would never be looked up, even though it
needs to be further propagated to the final consuming executable
or shared library. This also means that we will look up the Private
module in that case, but we can't do much about that.

Amends ad7b94e163ac5c3959a7e38d7f48536be288a187

Pick-to: 6.10
Change-Id: Id89ea5bf6de8cb343c2afc0a42fd758cf57aba54
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>