Fix handling of invisible Unicode marks in number parsing
CLDR says we shouldn't let them get in the way of parsing. Some
existing data includes them, in various combinations, and the signs
those wrap are alway either U+2212 MINUS SIGN or the ASCII
HYPHEN-MINUS and PLUS SIGN, which are handled by special cases in any
case. So simply skipping over the invisibles, when they are
encountered and not matched as part of some other CLDR-supplied token,
suffices.

In order to handle the invisibles (and any we see fit to skip in
future), turn the last few quirk cases of nextToken() into a switch
that deals with the invisibles and startsWith() checks of the other
quirks. The diff is best viewed with leading space changes ignored.

This fixes some test-cses created explicitly to expose this issue. It
also turns some previously-invalid test-cases into valid ones, where
I'd tested (as part of the Qt 6 fixing of misguided "single character"
assumptions) that using only the "first character" of a sign (where it
happened not to be the sign itself) was broken - but the first
character, in such cases, was in fact an invisible mark we should have
been ignoring anyway.

This does incur a small slow-down (a few percent on average, but less
than the standard deviation of the fractional speed changes). I think
it's worth it for the correctness.

Fixes: QTBUG-139922
Change-Id: I4e1772efa760744c1aca8209f1680d60c4e9e2e9
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>