QMetaObjectPrivate: refactor indexOfMethodRelative()
As requested in code review, use QMetaMethod::MethodType enum instead of
QMocConstants::MethodFlags, to specify which type to search for;
MethodFlags has many more irrelevant enumerators wrt. this function.

Make it a regular function instead of a template; take the type to
search for as a parameter.

Use a switch-statement instead of ternaries, for better readability.

Change-Id: I699b2a146019b9f4cd4fb5441de6cae151ecce78
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>