QCocoaTheme: Resolve folder icons via NSWorkspace
The GetIconRef API we're using for the other theme icons is deprecated,
but sadly doesn't have a modern equivalent yet, as UTType doesn't provide
predefined UTTypes for all of the same icons, and we don't want to rely
on internal identifiers like com.apple.macbookpro-16-2023-space-gray in
place of kComputerIcon.

However for folder icons we can use NSWorkspace to either get an icon
for a specific folder, or the generic folder icon via UTTypeFolder.

This also automatically gives us the right tint for the folder icons
on macOS 26, reflecting the tint the user has chosen.

We no longer treat open and closed folders as visually different, as
macOS natively doesn't seem to do this anywhere I could find, and there
is no API to get the modern folder look in both these states with the
tint color preserved.

As before, we treat the "Link" variant of File and Folder icons as the
same as their non-link variants.

Pick-to: 6.10 6.8 
Change-Id: If9c6ab4cffcf90b41b16adce338f494de4147629
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>