QConnManInterface: fix Coverity AUTO_CAUSES_COPY
Coverity complains that we're storing the result of
QDBusReply::error() (which returns by cref) in an auto variable, which
is causing a copy.

QDBusError is expensive to copy, so use the QDBusReply API as it was
intended: By calling error() multiple times. Catching the result in a
const QDBusError& variable is not going to improve the code further,
neither efficiency- nor readability-wise.

As a drive-by, port from iostreams-style to the preferred (in library
code) printf-style qWarning().

Amends 2f8a9694af51cf941928626a3a04b5d9b1dee65d.

Task-number: QTBUG-133938
Coverity-Id: 896015
Coverity-Id: 896016
Change-Id: Ic78a900a4ee676fdcc6547332e77c63d58cde08f
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>